requirements.txt
# Flask
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.3.3
blinker==1.6.2
click==8.1.3
colorama==0.4.6
flask==2.3.2
itsdangerous==2.1.2

# Flask login
flask_login==0.6.2

# MySQL connector
mysql-connector-python==8.0.33 
protobuf==3.20.3